requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
build==1.2.1
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2024.7.4
    # via
    #   -r requirements.in
    #   requests
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
coverage==7.6.1
    # via -r requirements.in
cryptography==43.0.0
    # via
    #   -r requirements.in
    #   jwcrypto
django==5.0.7
    # via
    #   -r requirements.in
    #   django-filter
    #   django-oauth-toolkit
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-filter==24.2
    # via -r requirements.in
django-oauth-toolkit==2.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
idna==3.7
    # via
    #   -r requirements.in
    #   requests
inflection==0.5.1
    # via
    #   -r requirements.in
    #   drf-yasg
jwcrypto==1.5.6
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
markdown==3.6
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
packaging==24.1
    # via
    #   -r requirements.in
    #   build
    #   drf-yasg
    #   gunicorn
pip-tools==7.4.1
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.1.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   drf-yasg
requests==2.32.3
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   jwcrypto
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   drf-yasg
urllib3==2.2.2
    # via
    #   -r requirements.in
    #   requests
wheel==0.43.0
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
